DONE Query Auth and Fragment

+++ src/resolvers/Mutation.js
@@ -44,23 +44,21 @@ const Mutation = {
       token: jwt.sign({ userId: user.id}, 'secret')
     }
   },
-  async deleteUser(parent, args, { prisma }, info) {
-    const UserExists = await prisma.exists.User({ id: args.id })
-
-    if (!UserExists) {
-      throw new Error('User not found')
-    }
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
 
     return prisma.mutation.deleteUser({
       where: {
-        id: args.id
+        id: userId
       }
     }, info)
   },
-  updateUser(parent, args, { prisma }, info) {
+  updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.updateUser({
       where: {
-        id: args.id
+        id: userId
       },
       data: args.data
     }, info)
@@ -81,14 +79,38 @@ const Mutation = {
       }
     }, info)
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!postExists) {
+      throw new Error('Unable to delete post')
+    }
+
     return prisma.mutation.deletePost({
       where: {
-        id: args.id
+        id: args.id,
       }
     }, info)
   },
-  updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: arg.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!postExists) {
+      throw new Error('Unable to update post')
+    }
+
     return prisma.mutation.updatePost({
       where: {
         id: args.id
@@ -96,13 +118,14 @@ const Mutation = {
       data: args.data
     }, info)
   },
-  createComment(parent, args, { prisma }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         },
         post: {
@@ -113,14 +136,35 @@ const Mutation = {
       }
     }, info)
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment( {
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) throw new Error('Unable delete Comment')
+
     return prisma.mutation.deleteComment({
       where: {
         id: args.id
       }
     }, info)
   },
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const commentExists = await prisma.exists.Comment( {
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) throw new Error('Unable delete Comment')
+
     return prisma.mutation.updateComment({
       where: {
         id: args.id
